Create GasolineReceipt.java
The original code from the lesson. 
public class GasolineReceipt
{
 public static void main( String[]args)
 {
  System.out.println("+--------------------------+");
  System.out.println("|                                        |"); 
  System.out.println("|       CORNER STORE        |"); 
  System.out.println("|                                        |"); 
  System.out.println("|  2015-03-29     04:38PM|"); 
  System.out.println("|                                        |"); 
  System.out.println("|  Gallons:         10.870      |"); 
  System.out.println("|  Price/gallon:   $ 2.089    |"); 
  System.out.println("|                                         |"); 
  System.out.println("|  Fuel total:    $ 22.71        |"); 
  System.out.println("|                                         |"); 
  System.out.println("+--------------------------+"); 
 }
}
The objective is the find out the gas prices from where you live and change the code to reflect your area's gas prices so that the output reflects the changes. The only difference is the numbers. I made the mistake of using Tab a few times and got bad input. From this lesson i learned not to use the tab key because the tab key adds invisible characters and can cause your code not to be outputting correctly.