Update PrintingChoices.java
/*
Learn Java the Hard Way. 
Exercise 3 : PrintingChoices
This is the original source code provided by the author 
before any changes are made to it. 
Here is the Study Drill for the lesson: 
1. Add a printing statement at the end of the code to display the sentence 
“I am learning Java the Hard Way!” But break it up like line 19 so that you 
only have two or three words in each String and plus signs between. Make 
sure the spaces get included!
*/