Update RudeQuestions.java
/*
Learn Java the Hard Way. Lesson 8: Storing the humans responses.
This is the source code provided by the author before any 
changes are made. 
There is the Study Drill for the lesson: 
1) Add a new variable of your choosing. Add another question. (It doesn’t have to be rude.) 
Let the human put an answer to your question into your new variable, 
and display the human’s answer on the screen afterward.
*/