Update RudeQuestions.java
/*
Learn Java the Hard Way. Lesson 8: Storing the humans responses.
This is the assignment after the changes were made per the 
directions of the Study Drill. 
There is the Study Drill for the lesson: 
1) Add a new variable of your choosing. Add another question. (It doesn’t have to be rude.) 
Let the human put an answer to your question into your new variable, 
and display the human’s answer on the screen afterward.
*/