Update README.md
Learn Java the Hard Way by Graham Mitchell is a Book with  a series of Lessons with Study Drills at the end of them. The user reads along and types out all of the code into a basic text editor like Notepad++ and then manually changes, compiles and runs the code per the direction of the Study Drills. 

Inside of this repository are of my lessons from LJTHW. 

All assignments were entered in as original source code from author and the changes i made afters. There are branches for every assignment so that the merge request and changes can be viewed.