Create SecretWord.java
Learn Java The Hard Way Lesson 18:If Statements with Strings
This is the original provided source code provided by the 
book before any changes were made. 

Study Drills: 

1. Pick a second secret word that will also work. Then add an OR to the first if statement 
so that typing EITHER word makes it print out “That’s correct!”. For example, 
if I wanted the new secret word to be “mogul”, I could change the if statement 
to look like this: if ( guess.equals(secret) || guess.equals("mogul") ) { Pick a different word for yours, though.