Create BMICategories.java
Learn Java the Hard Way Lesson 19 : Mutual Exclusion with Chains of If and Else
This is the original source code provided by the book before any changes were made. 

Study Drill: 

Instead of making the human enter their BMI directly, allow them to type in their
height and weight and compute the BMI for them.