Create HighLow.java
Learn Java the Hard Way Lesson 26 : A Number-Guessing Game
This is the original code provided by the book before any 
changes from the Study Drill were applied. 

Study Drill: 

This code should be shorter! Notice that lines 18 and 19 are 
identical to lines 22 and 23. This is silly; there’s no reason 
to write the same code in both if statements. Make it so there is 
only one copy of these two lines and put if after both if statements 
(but still inside the body of the while loop). Confirm that the program 
still works.