Create ThirtyDays.java
Learn Java The Hard Way 22 : Making Decisions with a Big Switch
This is the source code provided by the book before any of the 
changes from the Study Drill are applied. 

Study Drill: 

Remove some of the break statements in the first switch and 
add some println() statements to confirm that it will set monthName 
to one value, then another, and yet another until it finally gets stopped by a break.