Create RunningTotal.java
Learn Java The Hard Way Lesson 30:Adding Values One at a Time
This is the original source code provided by the book before 
any changes from the study drill were applied. 

Study Drill: 
Rewrite the code to use a while loop instead of a do-while loop. 
(It won’t compile at first.) Add/change something before the loop 
starts so that it compiles and make sure it still works. Then change 
it back to a do-while. (And leave in whatever you changed to make 
the while version compile.) 

2. When the human types in a zero to end the program, it prints 
“The total so far” and then prints “The final total” right after ward.
That looks awkward , so fix the code so that it skips printing 
“The total so far” if the loop is about to end. Make sure it 
still prints it the rest of the time, though.