Update PigDice.java
Learn Java the Hard Way 32:The Dice Game Called ‘Pig’
This is the code after the changes from the study drill were applied. 

Study Drill: 

Modify the code for the computer’s turn so that it will stop after 20 points
or when it has earned enough points to win the game.