Create ReceiptRevisitedException.java
Learn Java The Hard Way 40 : Programs that Write to Files
This is the original source code provided by the book 
before any of the changes from the Study Drills were 
applied.

Study Drill: 

Pick a price per gallon and put it into a double variable, 
then let the human enter in how many gallons of gas they want. 
Figure out the total cost. Then print the receipt to the file, 
but substitute their details so the file will look different 
depending on what they put in. I would remove the bars on the 
right side of the receipt, since making things line up would be hard.