Update ReceiptRevisitedException.java
Pick a price per gallon and put it into a double variable, 
then let the human enter in how many gallons of gas they want. 
Figure out the total cost. Then print the receipt to the file, 
but substitute their details so the file will look different 
depending on what they put in. I would remove the bars on the 
right side of the receipt, since making things line up would be hard.