Create CaesarCipher.java
Learn Java the Hard Way 45: Caesar Cipher (Looping ThroughaString)
This is the original code provided by the book before any changes 
from the study drill were applied. 

Study Drill: 

Modify the code to read the plain text message from a text file 
instead of letting the human type it in. Then, instead of just 
displaying the ciphertext on the screen, add code to store the 
ciphertext into a file.