Create NestingLoops.java
Learn Java The Hard Way Lesson Nested For Loops
This is the original source code provided by the book 
before any of the changes from the study drill were applied. 

Study Drill: 

1. Change the order of the loops so that the “c” loop is 
on the inside and the “n” loop is on the outside.
2. Add a System.out.println() statement where indicated 
(after the close brace of the inner loop (the “b” loop), 
but still inside the outer loop).