Update ArrayIntro.java
Learn Java The Hard Way 48 : Arrays - Many Values in a Single Variable
This is code after the changes from the study drill are applied. 

Study Drill: 

Create a second array of Strings and use an initializer list to put several 
values into it. Then add a second foreach loop to display the values from 
the new array on the screen, one at a time.