Create PickACard.java
Learn Java The Hard WAy Lesson Exercise 57 : A Deck of Playing Cards
This is the original source code from the book before any of the changes
from the study drill were applied.

Study Drill: 

Add a function called shuffleDeck(). It should take in an array of cards 
as a parameter and return nothing(void). One way to shuffle is to choose
two random numbers from 0-51 and “swap” the cards in those slots.Then put 
that code in a loop that repeats 1000 times or so. This is a bit tricky 
to get right.