Update ActorList.java
Learn Java The Hard Way Lesson 54 : Records from a File
This is the source code provided by the book 
after the changes from the study drill were applied. 

Study Drill: 

Add a counter inside the loop that reads in all the records. 
After the loop is over display a message on the screen about 
how many records were processed.