Create ArraySlotAccess.java
Learn Java The Hard Way Lesson 
This is the original provided source code before any of the changes from the 
Study Drill were applied. 

Study Drill: 

At the top of the code, change it so the array has a capacity of 1000 instead 
of 3. Don’t change any other code and recompile and run it again. Guess what? 
Those for loops might have been a little more work to write and to understand,
but once written they work just as well for 1000 values as for 3. And that’s 
pretty cool.