Update ArraySlotAccess.java
Learn Java The Hard Way Lesson 51: Arrays Without Foreach Loops

This is the source code after the changes from the 
Study Drill were applied. 

Study Drill: 

At the top of the code, change it so the array has a capacity of 1000 instead 
of 3. Don’t change any other code and recompile and run it again.