Create ItemNotFound.java
Learn Java The Hard Way Lesson 
This is the original source code provided by the book 
before any of the changes from the study drill were applied.

Study Drill: 

Either create a new array with different values or change some 
of the names in the array of Greek mortal heroes so that there 
are some duplicates. Then add a second for loop that uses a counter 
to determine if a value is found.