Update ItemNotFound.java
Learn Java The Hard Way Lesson 50 : Saying Something Is NOT in an Array
This is the source code after the changes from the study drill were applied.

Study Drill: 

Create a new array with different values or change some 
of the names in the array of Greek mortal heroes so that there 
are some duplicates. Then add a second for loop that uses a counter 
to determine if a value is found.