Create NestingLoops.java
Learn Java The Hard Way 46 : Nested For Loops
This is the original source code provided by the book before any of the 
changes from the Strudy Drill were applied. 

Study Drill: 

1)Change the order of the loops so that the “c” loop is on the 
inside and the “n” loop is on the outside. How does the output change? 
2) Look at the second set of nested loops (“AB”). Change the print() 
statement to println(). How does the output change? (Then change it back to print().)