Update README.md
Learn Java the Hard Way by Graham Mitchell is a Book with  a series of Lessons with Study Drills at the end of them. The user reads along and types out all of the code into a basic text editor like Notepad++ and then manually changes, compiles and runs the code per the direction of the Study Drills. 

Inside of this repository are of my lessons from LJTHW. 

The lessons from LJTHW cover:

*Installing Java, compiling and running Java files from cmd line. 
*Input and output.
*Boolean conditions, if statements, else if, else, == true, == false. 
*For Loops, While loops, Do while loops, Foreach loops. 
*Importing Java Libraries.
*Exception Handling.
*Arrays, Array Lists. 
*Methods, void methods, returning values, writing methods, calling methods. 
*Declaring and initializing variables, integers, doubles, Strings, Characters, Boolean. 
*Reading and outputting to files. 
*Counting text files and arrays using loops. 
*Comparing integers, doubles, strings. 

 There are a total of 58 lessons. I uploaded the lessons where i did make changes. Some of them required no changes to be
 made and therefore aren't listed. 

All assignments were entered in as original source code from author and the changes i made afters. There are branches for every assignment so that the merge request and changes can be viewed.